add LED spacer for enclosure considerations
I want the LED to spaced higher off the board so the converging lens in
the enclosure can be centered on the LED without the board having to be
so high up in the enclosure. This will make the PCB standoffs shorter
and give more space for the LCD and pushbutton wiring.