use part instead of assembly for pcb
For some reason, solidworks was not behaving properly with the pcb subassembly that referenced the step file. The assembly kept reloading the old step file without the led spacer/standoff.

Importing the step file with 3D interconnect turned off, and saving it as a multibody part seems to be working now; the pcb part doesn't get messed up when I close/reopen the main enclosure assembly, so now my mates aren't breaking :)